pipeline-short/requirements.txt
pandas 
pylint 
pytest
requests
aiohttp
python-dotenv
pymssql
